refactor: [#142] reorganize tests for the API
